create pull request action to open a PR in browser
